fix: data in-sync within the app
This client component will rerender the server component. This is done because earlier if a user tried adding an image to favorites, it would show up in the favorites list but, when the user revisited the gallery page, the heart icon of the image inside the gallery still remained the same as of previous.