feat: 🚧 search image by tags
Implement the search images by tags functionality.  User enters the tag name which is replaced inside the url when user submits the tag name. From there we grab that searchParam inside our gallery page as a prop. This searchParams object is used to add an expression for tags inside the search method from Search API.